Debugged boundary attribute assignments
